+++ Framework/Core/CashDispenser/Completions/Denominate_g.cs
@@ -26,7 +26,7 @@ namespace XFS4IoT.CashDispenser.Completions
         public sealed class PayloadData : MessagePayload
         {
 
-            public PayloadData(CompletionCodeEnum CompletionCode, string ErrorDescription, ErrorCodeEnum? ErrorCode = null, Dictionary<string, double> Currencies = null, Dictionary<string, double> Values = null, Dictionary<string, double> CashBox = null)
+            public PayloadData(CompletionCodeEnum CompletionCode, string ErrorDescription, ErrorCodeEnum? ErrorCode = null, Dictionary<string, double> Currencies = null, Dictionary<string, int> Values = null, Dictionary<string, double> CashBox = null)
                 : base(CompletionCode, ErrorDescription)
             {
                 this.ErrorCode = ErrorCode;
@@ -91,7 +91,7 @@ namespace XFS4IoT.CashDispenser.Completions
             /// command. The value of the entry is the number of items to take from that unit.
             /// </summary>
             [DataMember(Name = "values")]
-            public Dictionary<string, double> Values { get; init; }
+            public Dictionary<string, int> Values { get; init; }
 
             /// <summary>
             /// Only applies to Teller Dispensers. Amount to be paid from the teller’s cash box.
